Move `ast` -> `lib/ast`
