+++ src/lib/ast/get-dependencies-templates.js
@@ -1,6 +1,6 @@
 import nunjucks from 'nunjucks';
 
-import {ImportWrapper} from '../import-wrapper/ImportWrapper';
+import {ImportWrapper} from '../../import-wrapper/ImportWrapper';
 
 import {getAddNodeValue} from './get-add-node-value';
