+++ src/lib/ast/get-used-extensions.js
