+++ src/lib/ast/get-used-globals.js
